Rename c.sh -> c-home.sh since it has many computer-specific short-cuts.
First, rename c.sh to c-home.sh as it'll actually only be used on one
computer.

Second, remove the sourcing of c.sh entirely from cf/bashrc as this
bashrc is meant to be a file of common bash configurations and
machine-specific shell functions should be placed directly in ~/.bashrc.